Pulled the stat display from generated sets into its own component
