+++ src/app/loadout-builder/generated-sets/GeneratedSet.m.scss
@@ -11,17 +11,6 @@
   }
 }
 
-.light {
-  composes: light from global;
-  float: none;
-
-  :global(.app-icon) {
-    font-size: 10px;
-    vertical-align: -0.125em;
-    margin-right: 0;
-  }
-}
-
 .header {
   color: #ddd;
   display: flex;
@@ -34,16 +23,6 @@
     height: 16px;
     width: 16px;
   }
-}
-
-.statSegment {
-  &::after {
-    content: ' | ';
-  }
-}
-
-.nonActiveStat {
-  opacity: 0.4;
 }
 
 .items {
@@ -64,9 +43,4 @@
       margin-right: 0;
     }
   }
-}
-
-.warning {
-  color: yellow;
-  margin-right: 4px;
 }
