+++ src/app/loadout-builder/generated-sets/GeneratedSet.tsx
@@ -3,22 +3,15 @@ import { DimStore } from '../../inventory/store-types';
 import { ArmorSet, StatTypes, LockedMap, LockedArmor2ModMap } from '../types';
 import GeneratedSetButtons from './GeneratedSetButtons';
 import GeneratedSetItem from './GeneratedSetItem';
-import { powerIndicatorIcon, AppIcon, faExclamationTriangle } from '../../shell/icons';
 import _ from 'lodash';
-import { getNumValidSets, calculateTotalTier, sumEnabledStats } from './utils';
-import { statTier } from '../utils';
+import { getNumValidSets } from './utils';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
-import BungieImage from 'app/dim-ui/BungieImage';
-import { DestinyStatDefinition } from 'bungie-api-ts/destiny2';
-import { statHashes } from '../types';
-import { t } from 'app/i18next-t';
 import styles from './GeneratedSet.m.scss';
 import { editLoadout } from 'app/loadout/LoadoutDrawer';
 import { Loadout } from 'app/loadout/loadout-types';
 import { assignModsToArmorSet } from '../mod-utils';
-import { Armor2ModPlugCategories } from 'app/utils/item-utils';
 import { LoadoutBuilderAction } from '../loadoutBuilderReducer';
-import PressTip from 'app/dim-ui/PressTip';
+import SetStats from './SetStats';
 
 interface Props {
   set: ArmorSet;
@@ -60,96 +53,20 @@ function GeneratedSet({
     return null;
   }
 
-  const stats = _.mapValues(statHashes, (statHash) => defs.Stat.get(statHash));
-
-  const totalTier = calculateTotalTier(set.stats);
-  const enabledTier = sumEnabledStats(set.stats, enabledStats);
-
-  const displayStats = { ...set.stats };
-
-  // Add general mod vaues for display purposes
-  if ($featureFlags.armor2ModPicker) {
-    for (const lockedMod of lockedArmor2Mods[Armor2ModPlugCategories.general]) {
-      for (const stat of lockedMod.mod.investmentStats) {
-        if (stat.statTypeHash === statHashes.Mobility) {
-          displayStats.Mobility += stat.value;
-        } else if (stat.statTypeHash === statHashes.Recovery) {
-          displayStats.Recovery += stat.value;
-        } else if (stat.statTypeHash === statHashes.Resilience) {
-          displayStats.Resilience += stat.value;
-        } else if (stat.statTypeHash === statHashes.Intellect) {
-          displayStats.Intellect += stat.value;
-        } else if (stat.statTypeHash === statHashes.Discipline) {
-          displayStats.Discipline += stat.value;
-        } else if (stat.statTypeHash === statHashes.Strength) {
-          displayStats.Strength += stat.value;
-        }
-      }
-    }
-  }
-
   const assignedMods = $featureFlags.armor2ModPicker
     ? assignModsToArmorSet(set.firstValidSet, lockedArmor2Mods)
     : {};
 
-  const incorrectStats = _.uniq(
-    set.firstValidSet
-      .map((item) =>
-        item.stats
-          ?.filter((stat) => stat.statHash !== -1000)
-          .map((stat) => stat.baseMayBeWrong && stat.displayProperties.name)
-      )
-      .flat()
-      .filter(Boolean)
-  );
-
   return (
     <div className={styles.build} style={style} ref={forwardedRef}>
       <div className={styles.header}>
-        <div>
-          <span>
-            {set.firstValidSet.some((item) => item.stats?.some((stat) => stat.baseMayBeWrong)) && (
-              <PressTip
-                elementType="span"
-                tooltip={t('LoadoutBuilder.StatIncorrectWarning', {
-                  stats: incorrectStats.join('/'),
-                })}
-              >
-                <AppIcon className={styles.warning} icon={faExclamationTriangle} />
-              </PressTip>
-            )}
-            <span className={styles.statSegment}>
-              <span>
-                <b>
-                  {t('LoadoutBuilder.TierNumber', {
-                    tier: enabledTier,
-                  })}
-                </b>
-              </span>
-              {enabledTier !== totalTier && (
-                <span className={styles.nonActiveStat}>
-                  <b>
-                    {` (${t('LoadoutBuilder.TierNumber', {
-                      tier: totalTier,
-                    })})`}
-                  </b>
-                </span>
-              )}
-            </span>
-            {statOrder.map((stat) => (
-              <Stat
-                key={stat}
-                isActive={enabledStats.has(stat)}
-                stat={stats[stat]}
-                value={displayStats[stat]}
-              />
-            ))}
-          </span>
-          <span className={styles.light}>
-            <AppIcon icon={powerIndicatorIcon} /> {set.maxPower}
-          </span>
-        </div>
-
+        <SetStats
+          defs={defs}
+          set={set}
+          statOrder={statOrder}
+          enabledStats={enabledStats}
+          lockedArmor2Mods={lockedArmor2Mods}
+        />
         <GeneratedSetButtons
           numSets={numSets}
           set={set}
@@ -172,29 +89,6 @@ function GeneratedSet({
         ))}
       </div>
     </div>
-  );
-}
-
-function Stat({
-  stat,
-  isActive,
-  value,
-}: {
-  stat: DestinyStatDefinition;
-  isActive: boolean;
-  value: number;
-}) {
-  return (
-    <span
-      className={isActive ? styles.statSegment : `${styles.statSegment} ${styles.nonActiveStat}`}
-    >
-      <b>
-        {t('LoadoutBuilder.TierNumber', {
-          tier: statTier(value),
-        })}
-      </b>{' '}
-      <BungieImage src={stat.displayProperties.icon} /> {stat.displayProperties.name}
-    </span>
   );
 }
