+++ src/app/loadout-builder/generated-sets/SetStats.m.scss.d.ts
@@ -1,9 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'build': string;
-  'header': string;
-  'items': string;
+  'light': string;
+  'nonActiveStat': string;
+  'statSegment': string;
+  'warning': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
